Change and update database and models
Changed the database from Postgresql to SQLite. I
was initially using Postgresql to use the
ArrayField class, but since I'm not storing data,
using Postgresql does not seem necessary to set
up. SQLite is a more appropriate database to use
for this project.

I updated the ScrapedRecipe model to use JSONField
instead of ArrayField.